+++ variables.tf
@@ -172,7 +172,7 @@ variable "add_name_to_public_ip_addresses" {
   }
 }
 
-variable "enable_azure_bastion" {
+variable "add_bastion" {
   type        = bool
   default     = false
   description = "Specify if Azure Bastion Developer should be provisioned. See https://go.microsoft.com/fwlink/?linkid=2249215 for more information."